requirements.txt
chardet==3.0.4
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pdfminer.six==20181108
pycryptodome==3.7.3
six==1.12.0
sortedcontainers==2.1.0
Werkzeug==0.14.1
